+++ packages/browser/tests/unit/go.spec.ts
@@ -1,5 +1,5 @@
 import "jest";
-import { Browser, PUSH, REPLACE, POP } from "../../src";
+import { Browser } from "../../src";
 import { JSDOM } from "jsdom";
 
 function ignoreFirstCall(fn) {
@@ -58,10 +58,10 @@ describe("go", () => {
           cancelGo = pending.cancel;
           // trigger a push call and don't resolve the go
           testHistory.respondWith(pushRouter);
-          testHistory.navigate("/seven", PUSH);
+          testHistory.navigate("/seven", "push");
         });
         const pushRouter = ignoreFirstCall(function(pending) {
-          cancelGo(PUSH);
+          cancelGo("push");
           setTimeout(() => {
             expect(window.location.pathname).toBe("/six");
             done();
@@ -69,11 +69,11 @@ describe("go", () => {
         });
 
         testHistory.respondWith(initialRouter);
-        testHistory.navigate("/two", PUSH); // 1.0
-        testHistory.navigate("/three", PUSH); // 2.0
-        testHistory.navigate("/four", PUSH); // 3.0
-        testHistory.navigate("/five", PUSH); // 4.0
-        testHistory.navigate("/six", PUSH); // 5.0
+        testHistory.navigate("/two", "push"); // 1.0
+        testHistory.navigate("/three", "push"); // 2.0
+        testHistory.navigate("/four", "push"); // 3.0
+        testHistory.navigate("/five", "push"); // 4.0
+        testHistory.navigate("/six", "push"); // 5.0
 
         testHistory.respondWith(goRouter);
         testHistory.go(-2);
@@ -92,10 +92,10 @@ describe("go", () => {
           cancelGo = pending.cancel;
           // trigger a push call and don't resolve the go
           testHistory.respondWith(replaceRouter);
-          testHistory.navigate("/seven", REPLACE);
+          testHistory.navigate("/seven", "replace");
         });
         const replaceRouter = ignoreFirstCall(function(pending) {
-          cancelGo(REPLACE);
+          cancelGo("replace");
           setTimeout(() => {
             expect(window.location.pathname).toBe("/six");
             done();
@@ -103,11 +103,11 @@ describe("go", () => {
         });
 
         testHistory.respondWith(initialRouter);
-        testHistory.navigate("/two", PUSH); // 1.0
-        testHistory.navigate("/three", PUSH); // 2.0
-        testHistory.navigate("/four", PUSH); // 3.0
-        testHistory.navigate("/five", PUSH); // 4.0
-        testHistory.navigate("/six", PUSH); // 5.0
+        testHistory.navigate("/two", "push"); // 1.0
+        testHistory.navigate("/three", "push"); // 2.0
+        testHistory.navigate("/four", "push"); // 3.0
+        testHistory.navigate("/five", "push"); // 4.0
+        testHistory.navigate("/six", "push"); // 5.0
 
         testHistory.respondWith(goRouter);
         testHistory.go(-2);
@@ -129,7 +129,7 @@ describe("go", () => {
           testHistory.go(-1);
         });
         const popRouter = ignoreFirstCall(function(pending) {
-          cancelGo(POP);
+          cancelGo("pop");
           setTimeout(() => {
             expect(window.location.pathname).toBe("/three");
             done();
@@ -137,11 +137,11 @@ describe("go", () => {
         });
 
         testHistory.respondWith(initialRouter);
-        testHistory.navigate("/two", PUSH); // 1.0
-        testHistory.navigate("/three", PUSH); // 2.0
-        testHistory.navigate("/four", PUSH); // 3.0
-        testHistory.navigate("/five", PUSH); // 4.0
-        testHistory.navigate("/six", PUSH); // 5.0
+        testHistory.navigate("/two", "push"); // 1.0
+        testHistory.navigate("/three", "push"); // 2.0
+        testHistory.navigate("/four", "push"); // 3.0
+        testHistory.navigate("/five", "push"); // 4.0
+        testHistory.navigate("/six", "push"); // 5.0
 
         testHistory.respondWith(goRouter);
         testHistory.go(-2);
@@ -170,8 +170,8 @@ describe("go", () => {
       testHistory.confirmWith(confirm);
       testHistory.respondWith(router);
 
-      testHistory.navigate("/two", PUSH); // 1.0
-      testHistory.navigate("/three", PUSH); // 2.0
+      testHistory.navigate("/two", "push"); // 1.0
+      testHistory.navigate("/three", "push"); // 2.0
       setup = true;
       testHistory.go(-2);
     });
@@ -186,8 +186,8 @@ describe("go", () => {
       }
       testHistory.respondWith(router);
 
-      testHistory.navigate("/two", PUSH); // 1.0
-      testHistory.navigate("/three", PUSH); // 2.0
+      testHistory.navigate("/two", "push"); // 1.0
+      testHistory.navigate("/three", "push"); // 2.0
       // don't add function until we have setup the history
       testHistory.confirmWith(confirm);
       testHistory.go(-2);