+++ packages/hash/types/index.d.ts
@@ -1,4 +1,3 @@
-import { PUSH, REPLACE, ANCHOR, POP } from "@hickory/root";
 import { History, LocationDetails, HickoryLocation, PartialLocation, AnyLocation, Options as RootOptions } from "@hickory/root";
 export { History, HickoryLocation, PartialLocation, AnyLocation, LocationDetails };
 export interface Options extends RootOptions {
@@ -6,4 +5,4 @@ export interface Options extends RootOptions {
     hashType?: string;
 }
 declare function Hash(options?: Options): History;
-export { Hash, PUSH, REPLACE, ANCHOR, POP };
+export { Hash };