+++ packages/root/src/constants.ts
@@ -1,6 +1,0 @@
-import { PushType, ReplaceType, AnchorType, PopType } from "./types/hickory";
-
-export const PUSH: PushType = "push";
-export const REPLACE: ReplaceType = "replace";
-export const ANCHOR: AnchorType = "anchor";
-export const POP: PopType = "pop";