+++ packages/root/types/constants.d.ts
@@ -1,5 +1,0 @@
-import { PushType, ReplaceType, AnchorType, PopType } from "./types/hickory";
-export declare const PUSH: PushType;
-export declare const REPLACE: ReplaceType;
-export declare const ANCHOR: AnchorType;
-export declare const POP: PopType;