+++ packages/root/types/index.d.ts
@@ -1,4 +1,3 @@
 export * from "./types";
 import Common from "./common";
-import { PUSH, REPLACE, ANCHOR, POP } from "./constants";
-export { Common, PUSH, REPLACE, ANCHOR, POP };
+export { Common };